#리팩토링 db 정리
