+++ apollo/db/reviews.ts
@@ -1,19 +1,10 @@
-import mysql from "mysql2/promise";
-import { dbConfig } from "./db.config.js";
-const pool1 = mysql.createPool({
-  host: dbConfig.host,
-  user: dbConfig.user,
-  password: dbConfig.password,
-  database: dbConfig.database,
-  waitForConnections: true,
-  connectionLimit: 10,
-  queueLimit: 0,
-});
-export const review = async () => {
+import { pool1 } from "../mysqlPool.js";
+
+export const reviews = async () => {
   const [rows] = await pool1.query("select * from review");
   return rows;
 };
-export const getItemReview = async (id: number) => {
+export const reviewById = async (id: number) => {
   const [rows] = await pool1.query(
     `select * from review,user_master where review.user_code = user_master.user_code and review.sId=${id}`
   );
@@ -34,4 +25,3 @@ export const deleteReview = async (id: number) => {
   const [rows] = await pool1.query(`DELETE FROM review where rId=${id}`);
   return rows;
 };
-
