+++ apollo/mysqlPool.ts
@@ -1,0 +1,12 @@
+import mysql from "mysql2/promise";
+import { dbConfig } from "./db.config.js";
+
+export const pool1 = mysql.createPool({
+  host: dbConfig.host,
+  user: dbConfig.user,
+  password: dbConfig.password,
+  database: dbConfig.database,
+  waitForConnections: true,
+  connectionLimit: 10,
+  queueLimit: 0,
+});
